+++ apps/journeys-admin/src/components/Editor/JourneyEdit/JourneyEdit.tsx
@@ -5,10 +5,6 @@ import {
   ActiveJourneyEditContent
 } from '@core/journeys/ui/EditorProvider'
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
-import {
-  BlockFields_ButtonBlock as ButtonBlock,
-  BlockFields_ButtonBlock_action_LinkAction as LinkAction
-} from '../../../../__generated__/BlockFields'
 import { Canvas } from '../Canvas'
 import { ControlPanel } from '../ControlPanel'
 import { Drawer, DRAWER_WIDTH } from '../Drawer'
@@ -20,15 +16,6 @@ export function JourneyEdit(): ReactElement {
     state: { journeyEditContentComponent }
   } = useEditor()
   const { journey } = useJourney()
-
-  const actions = (journey?.blocks ?? [])
-    .filter((block) => ((block as ButtonBlock).action as LinkAction) != null)
-    .map((block) => (block as ButtonBlock).action as LinkAction)
-    .filter(
-      (action, i, arr) =>
-        ['LinkAction'].includes(action.__typename) &&
-        arr.findIndex((x) => x.url === action.url) === i
-    )
 
   return (
     <>
@@ -55,13 +42,13 @@ export function JourneyEdit(): ReactElement {
               {
                 [ActiveJourneyEditContent.Canvas]: <Canvas />,
                 [ActiveJourneyEditContent.Action]: (
-                  <ActionsTable actions={actions} />
+                  <ActionsTable journey={journey} />
                 )
               }[journeyEditContentComponent]
             }
           </Box>
         </Box>
-        <ControlPanel action={actions[0]} />
+        <ControlPanel />
       </Box>
       <Drawer />
     </>
